+++ CHANGELOG.md
@@ -37,6 +37,9 @@
 - Deprecate `pmetric.MetricAggregationTemporality` enum type in favor of `pmetric.AggregationTemporality` (#6249)
 - Deprecate `confmap.Conf.UnmarshalExact` in favor of `confmap.Conf.Unmarshal(.., WithErrorUnused)` (#6231)
 - Deprecate `pmetric.MetricDataPointFlags` favor of `pmetric.DataPointFlags` (#6259)
+- Deprecate `ptrace.SpanStatus` favor of `ptrace.Status` (#6258)
+- Deprecate `pmetric.[New]?Buckets` in favor of `pmetric.[New]?ExponentialHistogramDataPointBuckets` (#6261)
+- Deprecate `plog.SeverityNumberUndefined` in favor of `plog.SeverityNumberUnspecified` (#6269)
 
 ### 💡 Enhancements 💡